--- src/components/dashboard/ActiveCasesWidget.tsx
@@ -1,14 +1,0 @@
-
-import { StatCard } from "@/components/dashboard/StatCard";
-import { Briefcase } from "lucide-react";
-
-export function ActiveCasesWidget() {
-  return (
-    <StatCard
-      title="Active Cases"
-      value="0" 
-      icon={<Briefcase className="h-4 w-4 text-muted-foreground" />}
-      description="Currently ongoing cases"
-    />
-  );
-}
