--- src/components/dashboard/DraftsWidget.tsx
@@ -1,14 +1,0 @@
-
-import { StatCard } from "@/components/dashboard/StatCard";
-import { FileEdit } from "lucide-react";
-
-export function DraftsWidget() {
-  return (
-    <StatCard
-      title="Drafts"
-      value="0" 
-      icon={<FileEdit className="h-4 w-4 text-muted-foreground" />}
-      description="Cases saved as drafts"
-    />
-  );
-}
