--- src/components/dashboard/SearchBar.tsx
@@ -1,17 +1,0 @@
-import { Input } from "@/components/ui/input";
-import { Search } from "lucide-react";
-
-export function SearchBar() {
-  return (
-    <div className="relative w-full">
-      <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
-        <Search className="h-5 w-5 text-muted-foreground" />
-      </div>
-      <Input
-        type="search"
-        placeholder="Search case titles and chief complaints..."
-        className="pl-10 pr-4 py-2 w-full" // Added padding for the icon and some general styling
-      />
-    </div>
-  );
-}
